Add index on the timestamp in comments
There is an existing `idx_comment_timestamp` index but it is compound
on `(lbry_claim_id, timestamp). For recsys (and prob content moderation)
I need the ability to quickly query ALL recent queries.

Locally, this operation took 1m30 sec on my old (half-sized)
commentron backup. It should take less than 3 minutes on our server,
but the `comment` table is InnoDB so it shouldn't be a locking
operation.