Add comment classification and soft-deletes
This is a big commit that should have been done in stages
but was not. It has two major changes,

1. It adds soft deletes for comments instead of hard ones;
2. It adds a comment classification job to get a real-time view of behavior

(1) is an easy win by just using the built-in soft-deletes provided by
SQLBoiler.

To emphasis, (2) does no do auto-moderation. It only classifies comments
for human-in-the-loop interventions. It also still lacks one of the
most critical categories: doxxing.

In addition to these major changes it adds,

- A small docker script for testing against an ephemeral mysql instance
- Adds an execution guard to the OICD initialization so developers
  can work on the code without an internet connection without panics,
  set by `SKIP_OICD=1`